Extract buildEffectiveUseCaseMetadata to upload_metadata.ts
It composes upload-time metadata for all files, not just sandbox ones,
so it doesn't belong in sandbox_raw.ts. Sandbox predicates stay there.
