+++ front/lib/api/files/sandbox_raw.ts
@@ -1,9 +1,7 @@
-import { shouldSkipDataSourceIndexing } from "@app/lib/api/files/should_skip_indexing";
 import type { FileResource } from "@app/lib/resources/file_resource";
 import type {
   AllSupportedFileContentType,
   FileUseCase,
-  FileUseCaseMetadata,
 } from "@app/types/files";
 import { isSupportedDelimitedTextContentType } from "@app/types/files";
 
@@ -31,43 +29,4 @@ export function shouldStampSandboxRawDelimited({
     useCase === "conversation" &&
     isSupportedDelimitedTextContentType(contentType)
   );
-}
-
-export function buildEffectiveUseCaseMetadata({
-  contentType,
-  fileName,
-  flags,
-  providedMetadata,
-  useCase,
-}: {
-  contentType: AllSupportedFileContentType;
-  fileName: string;
-  flags: { hasSandboxTools: boolean };
-  providedMetadata: FileUseCaseMetadata | undefined;
-  useCase: FileUseCase;
-}): FileUseCaseMetadata | undefined {
-  const skipDataSourceIndexing = shouldSkipDataSourceIndexing({
-    contentType,
-    fileName,
-  });
-  const isSandboxRawDelimited = shouldStampSandboxRawDelimited({
-    contentType,
-    flags,
-    useCase,
-  });
-
-  if (!skipDataSourceIndexing && !isSandboxRawDelimited) {
-    return providedMetadata;
-  }
-
-  return {
-    ...(providedMetadata ?? {}),
-    ...(skipDataSourceIndexing ? { skipDataSourceIndexing: true } : {}),
-    ...(isSandboxRawDelimited
-      ? {
-          skipDataSourceIndexing: true,
-          skipFileProcessing: true,
-        }
-      : {}),
-  };
 }
